# agent/openapi/week2.yaml
components:
  schemas:
...
    Body_upload_document_v1_documents_post:
      properties:
        clinician_id:
          exclusiveMinimum: 0.0
          title: Clinician Id
          type: integer
        doc_type:
          default: lab_pdf
          title: Doc Type
          type: string
        file:
          description: The scanned lab PDF / intake form.
          format: binary
          title: File
          type: string
        patient_id:
          exclusiveMinimum: 0.0
          title: Patient Id
          type: integer
      required:
      - file
      - patient_id
      - clinician_id
      title: Body_upload_document_v1_documents_post
      type: object
    ChatRequest:
      description: 'One grounded question about a patient, optionally continuing a
        thread.


        ``clinician_id`` is optional: in ``disabled`` mode it identifies the acting

        clinician (as today); in ``smart`` mode the session cookie is authoritative

        and this field, if present, is only validated against it (mismatch → 403).'
      properties:
        clinician_id:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - type: 'null'
          title: Clinician Id
        conversation_id:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - type: 'null'
          title: Conversation Id
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        message:
          minLength: 1
          title: Message
          type: string
        patient_id:
          exclusiveMinimum: 0.0
          title: Patient Id
          type: integer
      required:
      - patient_id
      - message
      title: ChatRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: Clinical Co-Pilot
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/chat:
    post:
      operationId: chat_v1_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Chat V1 Chat Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Answer a grounded question about a patient
      tags:
      - chat
  /v1/conversations/{conversation_id}:
    get:
      operationId: get_conversation_v1_conversations__conversation_id__get
      parameters:
      - in: path
        name: conversation_id
        required: true
        schema:
          title: Conversation Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Conversation V1 Conversations  Conversation Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Read a conversation's turns in order
      tags:
      - chat
  /v1/documents:
    post:
      operationId: upload_document_v1_documents_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_v1_documents_post'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Upload Document V1 Documents Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload a source document and ingest it (async envelope)
      tags:
      - documents
  /v1/documents/{document_id}:
    get:
      operationId: get_document_v1_documents__document_id__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Document V1 Documents  Document Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Ingestion status, facts, and citations
      tags:
      - documents
  /v1/documents/{document_id}/pages/{page_no}:
    get:
      operationId: get_document_page_v1_documents__document_id__pages__page_no__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: integer
      - in: path
        name: page_no
        required: true
        schema:
          title: Page No
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: The rendered page image (bbox-overlay backdrop)
      tags:
      - documents
...